requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
py-mini-racer==0.6.0
ffmpeg-python==0.2.0
future==1.0.0
yt-dlp>=2023.11.16
pydantic==2.11.7
mcp>=1.0.0
